Flatten and reorganize Python packages
Instead of applications all being subpackages of "kolibri_gnome", create
more packages with clearer names, as well as a "kolibri_app" package
with common utilities. This change should make it easier to understand
the project.
