+++ data/applications/meson.build
@@ -4,7 +4,7 @@ desktop_file = i18n.merge_file(
     input: configure_file(
         input: 'org.learningequality.Kolibri.Frontend.desktop.in.in',
         output: 'org.learningequality.Kolibri.Frontend.desktop.in',
-        configuration: kolibri_gnome_config
+        configuration: kolibri_app_config
     ),
     output: frontend_application_id + '.desktop',
     type: 'desktop',
@@ -16,7 +16,7 @@ desktop_file = i18n.merge_file(
 launcher_file = configure_file(
   input: 'org.learningequality.Kolibri.Launcher.desktop.in',
   output: launcher_application_id + '.desktop',
-  configuration: kolibri_gnome_config,
+  configuration: kolibri_app_config,
   install: true,
   install_dir: applications_dir
 )
