+++ data/applications/org.learningequality.Kolibri.Frontend.desktop.in.in
@@ -8,7 +8,7 @@ Comment=Offline education technology platform
 #              This list must end with a semicolon.
 Keywords=education;multimedia;videos;learning;quiz;
 Categories=Education;
-Exec=kolibri-gnome %U
+Exec=kolibri-gnome-frontend %U
 StartupNotify=true
 DBusActivatable=true
 # TRANSLATORS: This is an icon file name.
@@ -16,4 +16,3 @@ DBusActivatable=true
 Icon=@FRONTEND_APPLICATION_ID@
 MimeType=x-scheme-handler/kolibri;x-scheme-handler/x-kolibri-app;
 X-Endless-LaunchMaximized=true
-
