+++ data/dbus/org.learningequality.Kolibri.Daemon.service.in
@@ -1,3 +1,3 @@
 [D-BUS Service]
 Name=@DAEMON_APPLICATION_ID@
-Exec=@KOLIBRI_GNOME_LIBEXECDIR@/kolibri-daemon --session
+Exec=@KOLIBRI_APP_LIBEXECDIR@/kolibri-daemon --session
