+++ data/dbus/org.learningequality.Kolibri.Frontend.service.in
@@ -1,3 +1,3 @@
 [D-BUS Service]
 Name=@FRONTEND_APPLICATION_ID@
-Exec=kolibri-gnome
+Exec=kolibri-gnome-frontend
