+++ data/dbus/org.learningequality.Kolibri.Launcher.service.in
@@ -1,3 +1,3 @@
 [D-BUS Service]
 Name=@LAUNCHER_APPLICATION_ID@
-Exec=kolibri-launcher
+Exec=@KOLIBRI_APP_LIBEXECDIR@/kolibri-gnome-launcher
