+++ data/dbus/org.learningequality.Kolibri.SearchProvider.service.in
@@ -1,3 +1,3 @@
 [D-BUS Service]
 Name=@SEARCH_PROVIDER_APPLICATION_ID@
-Exec=@KOLIBRI_GNOME_LIBEXECDIR@/kolibri-gnome-search-provider
+Exec=@KOLIBRI_APP_LIBEXECDIR@/kolibri-gnome-search-provider
