+++ data/gnome-shell/meson.build
@@ -3,6 +3,6 @@ gnome_shell_search_providers_dir = join_paths(datadir, 'gnome-shell', 'search-pr
 configure_file(
     input: 'org.learningequality.Kolibri.ini.in',
     output: base_application_id + '.ini',
-    configuration: kolibri_gnome_config,
+    configuration: kolibri_app_config,
     install_dir: gnome_shell_search_providers_dir
 )
