+++ data/kolibri-gnome/meson.build
@@ -1,1 +1,1 @@
-install_subdir('assets', install_dir: kolibri_gnome_datadir)
+install_subdir('assets', install_dir: kolibri_app_datadir)
