+++ meson.build
@@ -42,29 +42,29 @@ search_provider_object_path = base_object_path + '/SearchProvider'
 
 po_dir = join_paths(meson.source_root(), 'po')
 
-kolibri_gnome_libexecdir = join_paths(libexecdir, 'kolibri-gnome')
-kolibri_gnome_datadir = join_paths(datadir, 'kolibri-gnome')
+kolibri_app_libexecdir = join_paths(libexecdir, 'kolibri-app')
+kolibri_app_datadir = join_paths(datadir, 'kolibri-app')
 
-kolibri_gnome_config = configuration_data()
-kolibri_gnome_config.set('BINDIR', bindir)
-kolibri_gnome_config.set('PYTHON', 'python3')
-kolibri_gnome_config.set('PACKAGE_STRING', package_string)
-kolibri_gnome_config.set('GETTEXT_PACKAGE', meson.project_name())
-kolibri_gnome_config.set('LOCALE_DIR', locale_dir)
-kolibri_gnome_config.set('BUILD_PROFILE', build_profile)
-kolibri_gnome_config.set('KOLIBRI_GNOME_LIBEXECDIR', kolibri_gnome_libexecdir)
-kolibri_gnome_config.set('KOLIBRI_GNOME_DATADIR', kolibri_gnome_datadir)
-kolibri_gnome_config.set('KOLIBRI_HOME_TEMPLATE_DIR', kolibri_home_template_dir)
-kolibri_gnome_config.set('BASE_APPLICATION_ID', base_application_id)
-kolibri_gnome_config.set('BASE_OBJECT_PATH', base_object_path)
-kolibri_gnome_config.set('FRONTEND_APPLICATION_ID', frontend_application_id)
-kolibri_gnome_config.set('LAUNCHER_APPLICATION_ID', launcher_application_id)
-kolibri_gnome_config.set('FRONTEND_OBJECT_PATH', frontend_object_path)
-kolibri_gnome_config.set('FRONTEND_CHANNEL_APPLICATION_ID_PREFIX', frontend_channel_application_id_prefix)
-kolibri_gnome_config.set('DAEMON_APPLICATION_ID', daemon_application_id)
-kolibri_gnome_config.set('DAEMON_OBJECT_PATH', daemon_object_path)
-kolibri_gnome_config.set('SEARCH_PROVIDER_APPLICATION_ID', search_provider_application_id)
-kolibri_gnome_config.set('SEARCH_PROVIDER_OBJECT_PATH', search_provider_object_path)
+kolibri_app_config = configuration_data()
+kolibri_app_config.set('BINDIR', bindir)
+kolibri_app_config.set('PYTHON', 'python3')
+kolibri_app_config.set('PACKAGE_STRING', package_string)
+kolibri_app_config.set('GETTEXT_PACKAGE', meson.project_name())
+kolibri_app_config.set('LOCALE_DIR', locale_dir)
+kolibri_app_config.set('BUILD_PROFILE', build_profile)
+kolibri_app_config.set('KOLIBRI_APP_LIBEXECDIR', kolibri_app_libexecdir)
+kolibri_app_config.set('KOLIBRI_APP_DATADIR', kolibri_app_datadir)
+kolibri_app_config.set('KOLIBRI_HOME_TEMPLATE_DIR', kolibri_home_template_dir)
+kolibri_app_config.set('BASE_APPLICATION_ID', base_application_id)
+kolibri_app_config.set('BASE_OBJECT_PATH', base_object_path)
+kolibri_app_config.set('FRONTEND_APPLICATION_ID', frontend_application_id)
+kolibri_app_config.set('LAUNCHER_APPLICATION_ID', launcher_application_id)
+kolibri_app_config.set('FRONTEND_OBJECT_PATH', frontend_object_path)
+kolibri_app_config.set('FRONTEND_CHANNEL_APPLICATION_ID_PREFIX', frontend_channel_application_id_prefix)
+kolibri_app_config.set('DAEMON_APPLICATION_ID', daemon_application_id)
+kolibri_app_config.set('DAEMON_OBJECT_PATH', daemon_object_path)
+kolibri_app_config.set('SEARCH_PROVIDER_APPLICATION_ID', search_provider_application_id)
+kolibri_app_config.set('SEARCH_PROVIDER_OBJECT_PATH', search_provider_object_path)
 
 subdir('data')
 subdir('po')
