--- src/kolibri-daemon.in
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-@PYTHON@ -m "kolibri_gnome.kolibri_daemon.main" $@
