--- src/kolibri-gnome.in
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-@PYTHON@ -m "kolibri_gnome.desktop.main" $@
