--- src/kolibri-search-provider.in
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-@PYTHON@ -m "kolibri_gnome.search_provider.main" $@
