+++ src/kolibri_app/config.py.in
@@ -7,7 +7,7 @@ FRONTEND_OBJECT_PATH = "@FRONTEND_OBJECT_PATH@"
 FRONTEND_CHANNEL_APPLICATION_ID_PREFIX = "@FRONTEND_CHANNEL_APPLICATION_ID_PREFIX@"
 SEARCH_PROVIDER_APPLICATION_ID = "@SEARCH_PROVIDER_APPLICATION_ID@"
 SEARCH_PROVIDER_OBJECT_PATH = "@SEARCH_PROVIDER_OBJECT_PATH@"
-DATA_DIR = "@KOLIBRI_GNOME_DATADIR@"
+DATA_DIR = "@KOLIBRI_APP_DATADIR@"
 GETTEXT_PACKAGE = "@GETTEXT_PACKAGE@"
 KOLIBRI_HOME_TEMPLATE_DIR = "@KOLIBRI_HOME_TEMPLATE_DIR@"
 LOCALE_DIR = "@LOCALE_DIR@"
