+++ src/kolibri_daemon/kolibri-daemon.in
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+@PYTHON@ -m "kolibri_daemon.main" $@
