+++ src/kolibri_daemon/kolibri_service_main.py
@@ -3,9 +3,10 @@ import os
 from contextlib import contextmanager
 from pathlib import Path
 
-from ..globals import init_kolibri
-from ..globals import init_logging
-from ..globals import KOLIBRI_HOME_PATH
+from kolibri_app.globals import init_kolibri
+from kolibri_app.globals import init_logging
+from kolibri_app.globals import KOLIBRI_HOME_PATH
+
 from .content_extensions import ContentExtensionsList
 
 # TODO: We need to use multiprocessing because Kolibri occasionally calls
