+++ src/kolibri_gnome/kolibri-gnome.in
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+@PYTHON@ -m "kolibri_gnome.main" $@
