+++ src/kolibri_gnome_launcher/kolibri-gnome-launcher.in
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+@PYTHON@ -m "kolibri_gnome_launcher.main" $@
